add config in the release test
